requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
importlib-metadata==1.7.0
jdcal==1.4.1
Markdown==3.2.2
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.1
path==15.0.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
Unipath==1.1
zipp==3.1.0
